+++ test/bootstrap_other_components_test.rb
@@ -10,7 +10,7 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
 
     expected = <<-HTML.strip_heredoc
       <div class="form-group row">
-        <label class="form-control-label col-sm-2 required" for="user_email">Email</label>
+        <label class="col-sm-2 required" for="user_email">Email</label>
         <div class="col-sm-10">
           <p class="form-control-static">steve@example.com</p>
         </div>
@@ -26,7 +26,7 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
 
     expected = <<-HTML.strip_heredoc
       <div class="form-group row">
-        <label class="form-control-label col-sm-2" for="user_">My Label</label>
+        <label class="col-sm-2" for="user_">My Label</label>
         <div class="col-sm-10">
           <p class="form-control-static">this is a test</p>
         </div>
@@ -42,7 +42,7 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
 
     expected = <<-HTML.strip_heredoc
       <div class="form-group row">
-        <label class="form-control-label col-sm-2" for="user_">Custom Label</label>
+        <label class="col-sm-2" for="user_">Custom Label</label>
         <div class="col-sm-10">
           <p class="form-control-static">Custom Control</p>
         </div>
@@ -58,7 +58,7 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
 
     expected = <<-HTML.strip_heredoc
       <div class="form-group row">
-        <label class="form-control-label col-sm-2 required" for="user_email">Email</label>
+        <label class="col-sm-2 required" for="user_email">Email</label>
         <div class="col-sm-10">this is a test</div>
       </div>
     HTML
@@ -72,7 +72,7 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
 
     expected = <<-HTML.strip_heredoc
       <div class="form-group row">
-        <label class="form-control-label col-sm-2" for="user_">My Label</label>
+        <label class="col-sm-2" for="user_">My Label</label>
         <div class="col-sm-10">this is a test</div>
       </div>
     HTML
@@ -86,7 +86,7 @@ class BootstrapOtherComponentsTest < ActionView::TestCase
 
     expected = <<-HTML.strip_heredoc
       <div class="form-group row">
-        <label class="form-control-label col-sm-2" for="user_">Custom Label</label>
+        <label class="col-sm-2" for="user_">Custom Label</label>
         <div class="col-sm-10">Custom Control</div>
       </div>
     HTML