+++ CHANGELOG.md
@@ -2,11 +2,15 @@
 
 Bugfixes:
   - Your contribution here!
+  - form-control-danger is replaced with is-invalid for bootstrap 4.0.0.beta3
+  - form-control-feedback is replaced with invalid-feedback for bootstrap 4.0.0.beta3
+  - help texts are rendered with <small> tag instead of <span> tag, i.e. like in bootstrap 4.0.0.beta3 
 
 Features:
   - Your contribution here!
+  - new `custom: true` option for radio buttons and check boxes according to bootstrap 4.0.0.beta3
+  - Allow HTML in help translations by using the '_html' suffix on the key - [@unikitty37](https://github.com/unikitty37)
   * [#325](https://github.com/bootstrap-ruby/rails-bootstrap-forms/pull/325): Support :prepend and :append for the `select` helper - [@donv](https://github.com/donv).
-
 
 ## [2.7.0][] (2017-04-21)